♻️ Migrate `__tests__/` pages to TS
