+++ packages/website/package.json
@@ -14,6 +14,7 @@
     "test": "FORCE_COLOR=1 jest --coverage"
   },
   "devDependencies": {
+    "@types/fetch-mock": "^7.3.5",
     "@types/jest": "^29.2.4",
     "@types/react": "^18.0.26",
     "@types/react-test-renderer": "^18.0.0",
@@ -36,6 +37,7 @@
     "next-pwa": "^5.4.4",
     "next-sitemap": "^3.1.43",
     "next-themes": "^0.2.0",
+    "node-mocks-http": "^1.12.1",
     "prettier": "2.8.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
